requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
cffi==1.17.1
    # via cryptography
cryptography==44.0.0
    # via -r requirements.in
django==5.1.4
    # via -r requirements.in
pillow==11.1.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pymysql==1.1.1
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.5.3
    # via django
